+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from "./core/loader";
-export * from "./core/helpers";
-export * from "./types";
+export { loadEnv } from "./core/loader";
+export { EnvLoader, getEnv, getKeys, hasKeys } from "./core/helpers";
+export type { LoadOptions } from "./types";